% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfun.R
\name{polyfun}
\alias{polyfun}
\title{Polynomial function}
\usage{
polyfun(p)
}
\arguments{
\item{p}{Coefficients of the polynomial}
}
\value{
A function that evaluates the polynomial in given points.
}
\description{
Polynomial function
}
